Added more complete instrumentation for a variety of classes

+++ run.sh
@@ -4,7 +4,7 @@ set -eu
 scalaver="2.11"
 sparkver="2.3.0-SNAPSHOT" #$(cat $basedir/docs/_config.yml | grep '^SPARK_VERSION:' | cut -d ' ' -f 2)
 hadoopver="2.8.1"
-hadoopvershort=$(echo "$hadoopver" | cut -d '.' -f -2)
+hadoopvershort=2.7 #$(echo "$hadoopver" | cut -d '.' -f -2)
 
 basedir="$HOME/code/spark"
 distdir="$basedir/mindist"
@@ -18,7 +18,9 @@ master="${cluster}1.$domain"
 slaves="${cluster}{2..$nnodes}.$domain"
 dest="/test/spark-tracing"
 traceout="/tmp/spark-trace"
+localhadoop="/opt/hadoop"
 port="5010"
+local=1
 
 testclass="org.apache.spark.examples.SparkPi"
 testjar="spark-examples_${scalaver}-$sparkver.jar"
@@ -26,15 +28,27 @@ iters=20
 
 while [[ $# > 0 ]]
 	do case "$1" in
-	#"spark-dist")
-	#	"$basedir/dev/make-distribution.sh" --name spark-tracing -Pyarn -Phive -Phadoop-$hadoopvershort -Dhadoop.version=$hadoopver
-	#	;;
-	#"spark-clean")
-	#	"$basedir/build/mvn" -Phive -Pyarn -Phadoop-$hadoopvershort -Dhadoop.version=$hadoopver -DskipTests clean
-	#	;;
-	#"spark-build")
-	#	"$basedir/build/mvn" -Phive -Pyarn -Phadoop-$hadoopvershort -Dhadoop.version=$hadoopver -DskipTests package
-	#	;;
+	"spark-dist")
+		pushd "$basedir"
+		"dev/make-distribution.sh" --name spark-tracing -Pyarn -Phive -Phadoop-$hadoopvershort -Dhadoop.version=$hadoopver
+		popd
+		;;
+	"spark-clean")
+		pushd "$basedir"
+		"build/mvn" -Phive -Pyarn -Phadoop-$hadoopvershort -Dhadoop.version=$hadoopver -DskipTests clean
+		popd
+		;;
+	"spark-build")
+		pushd "$basedir"
+		"build/mvn" -Phive -Pyarn -Phadoop-$hadoopvershort -Dhadoop.version=$hadoopver -DskipTests package
+		popd
+		;;
+	"local")
+		local=1
+		;;
+	"remote")
+		local=0
+		;;
 	"clean")
 		pushd "instrument"
 		sbt clean
@@ -46,13 +60,15 @@ while [[ $# > 0 ]]
 		popd
 		;;
 	"conf")
+		rdest="$dest"
+		[[ "$local" = "1" ]] && rdest="$distdir"
 		cat <<- ! > $basedir/conf/spark-defaults.conf
 		spark.master.ui.port $port
 		spark.worker.ui.port $port
 		spark.hadoop.yarn.timeline-service.enabled false
 		spark.executor.instances 1
 		spark.executor.memory 512m
-		spark.yarn.jars local:$dest/jars/*
+		spark.yarn.jars local:$rdest/jars/*
 
 		spark.dynamicAllocation.cachedExecutorIdleTimeout 60s
 		spark.dynamicAllocation.enabled true
@@ -68,16 +84,16 @@ while [[ $# > 0 ]]
 		spark.executor.cores 1
 		spark.task.cpus 1
 
-		#spark.driver.extraJavaOptions -javaagent:$dest/instrument_2.11-1.0.jar
-		#spark.yarn.am.extraJavaOptions -javaagent:$dest/instrument_2.11-1.0.jar
-		spark.executor.extraJavaOptions -javaagent:$dest/instrument_2.11-1.0.jar
+		spark.driver.extraJavaOptions -javaagent:$rdest/instrument_2.11-1.0.jar -Diop.version=4.3.0.0
+		spark.yarn.am.extraJavaOptions -javaagent:$rdest/instrument_2.11-1.0.jar -Diop.version=4.3.0.0
+		spark.executor.extraJavaOptions -javaagent:$rdest/instrument_2.11-1.0.jar -Diop.version=4.3.0.0
 		!
 		cp $basedir/conf/log4j.properties{.template,}
 		cat <<- ! >> $basedir/conf/log4j.properties
 		log4j.rootCategory=INFO, console
 		!
 		cat <<- ! > $basedir/conf/spark-env.sh
-		SPARK_HOME=$dest
+		SPARK_HOME=$rdest
 		HADOOP_CONF_DIR=/etc/hadoop/conf
 		!
 
@@ -86,40 +102,34 @@ while [[ $# > 0 ]]
 		ln -s ../../common/network-yarn/target/scala-$scalaver/spark-${sparkver}-yarn-shuffle.jar $distdir/yarn
 		;;
 	"upload")
+		[[ "$local" = "1" ]] && continue
 		rsync -av --copy-unsafe-links --delete --progress $distdir/ $user@$master:$dest
-		ssh -t $user@$master "cat /etc/spark2/conf/spark-defaults.conf | sed '/spark.yarn.archive/s/^/#/' >> $dest/conf/spark-defaults.conf"
 		ssh -t $user@$master "for host in $slaves; do rsync -av --copy-unsafe-links --delete $dest/ \$host:$dest; done"
 		;;
-	"remote")
-		ssh -t $user@$master "sudo rm -rf $traceout; for host in $slaves; do ssh -t \$host sudo rm -rf $traceout; done"
-		#ssh -t $user@$master "$dest/sbin/stop-all.sh; for host in $slaves; do ssh \$host $dest/sbin/stop-all.sh; done"
-		#ssh -t $user@$master "$dest/sbin/start-master.sh; for host in $slaves; do ssh \$host $dest/sbin/start-slave.sh $master:7077; done"
-		#firefox http://$master:$port &
-		ssh -t $user@$master sudo -iu notebook SPARK_PRINT_LAUNCH_COMMAND=1 $dest/bin/spark-submit --conf spark.driver.extraJavaOptions="-javaagent:$dest/instrument_2.11-1.0.jar" --conf spark.yarn.am.extraJavaOptions="-javaagent:$dest/instrument_2.11-1.0.jar" --master yarn --class "$testclass" "$dest/examples/jars/$testjar" $iters
+	"run")
+		if [[ "$local" = "0" ]]; then
+			ssh -t $user@$master "sudo rm -rf $traceout; for host in $slaves; do ssh -t \$host sudo rm -rf $traceout; done"
+			ssh -t $user@$master sudo -iu notebook $dest/bin/spark-submit --master yarn --class "$testclass" "$dest/examples/jars/$testjar" $iters
+		else
+			rm -rf $traceout
+			$distdir/bin/spark-submit --master yarn --class "$testclass" "$distdir/examples/jars/$testjar" $iters
+		fi
 		;;
 	"collect")
-		ssh -t $user@$master "sudo chown -R dev-user $traceout; for host in $slaves; do ssh -t \$host sudo chown -R dev-user $traceout; done"
-		ssh -t $user@$master "for host in $slaves; do rsync -av \$host:$traceout/ $traceout; done"
+		[[ "$local" = "1" ]] && continue
+		ssh -t $user@$master "sudo chown -R dev-user $traceout; for host in $slaves; do ssh -t \$host sudo chown -R dev-user $traceout; done" || true
+		ssh -t $user@$master "for host in $slaves; do rsync -av \$host:$traceout/ $traceout; done" || true
 		localout="$resultdir/remote"
 		rm -rf "$localout"
 		mkdir "$localout"
 		rsync -av --progress $user@$master:$traceout/ "$localout"
-		ssh -t $user@$master "rm -r $traceout; for host in $slaves; do ssh -t \$host rm -r $traceout; done"
+		ssh -t $user@$master "rm -r $traceout; for host in $slaves; do ssh -t \$host rm -r $traceout; done" || true
 		;;
-	"local")
-		#$distdir/sbin/stop-all.sh
-		rm -rf $traceout
-		$distdir/bin/spark-submit --master yarn --class "$testclass" "$distdir/examples/jars/$testjar" $iters
-		;;
-	"standalone")
-		$distdir/sbin/stop-all.sh
-		rm -rf $traceout
-		#$distdir/sbin/start-all.sh
-		$distdir/sbin/start-master.sh
-		$distdir/sbin/start-slave.sh spark://localhost:7077
-
-		#firefox http://localhost:$port &
-		$distdir/bin/spark-submit --master spark://localhost:7077 --class "$testclass" "$distdir/examples/jars/$testjar" $iters
+	"yarn")
+		[[ "$local" = "0" ]] && continue
+		"$localhadoop/sbin/stop-yarn.sh" || true
+		YARN_USER_CLASSPATH="$distdir/yarn/spark-$sparkver-yarn-shuffle.jar" "$localhadoop/sbin/yarn-daemon.sh" start resourcemanager
+		YARN_USER_CLASSPATH="$distdir/yarn/spark-$sparkver-yarn-shuffle.jar" "$localhadoop/sbin/yarn-daemon.sh" start nodemanager
 		;;
 	*)
 		echo "Unknown action $1"
